Lots of Changes

+++ user/user.go
@@ -36,10 +36,10 @@ func Authenticate(c *fiber.Ctx) error {
 	db := database.DBConn
 	user := new(User)
 	message := fiberutils.Message{}
-	err := fiberutils.ParseBody(c, user)
+	err := fiberutils.ParseBody(c, user) //(c (pagkukuhanan), user (pagpapasahan))
 	username := user.Username
 	password := user.Password
-	db.First(user, "username = ?", user.Username)
+	db.First(user, "username = ?", user.Username) //(user , "username = ?" (?-placeholder), user.Username)
 
 	if passwordhashing.CheckPasswordHash(password, user.Password) {
 		message = fiberutils.Message{
@@ -48,12 +48,12 @@ func Authenticate(c *fiber.Ctx) error {
 		}
 
 		// Create token
-		token := jwt.New(jwt.SigningMethodHS256)
+		token := jwt.New(jwt.SigningMethodHS256) //generate a random hash token when user login, token usually used to identify or use to as a key to access some form
 
 		// Set claims
 		claims := token.Claims.(jwt.MapClaims)
-		claims["username"] = user.Username
-		claims["exp"] = time.Now().Add(time.Hour * 72).Unix()
+		claims["user"] = user
+		claims["exp"] = time.Now().Add(time.Minute * 15).Unix()
 
 		if user.Role == "admin" {
 			claims["admin"] = true
@@ -73,7 +73,7 @@ func Authenticate(c *fiber.Ctx) error {
 		c.Cookie(&fiber.Cookie{
 			Name:   "token",
 			Value:  t,
-			MaxAge: 15 * 60,
+			MaxAge: 15 * 60, //(15 (minutes) * 60 (seconds))
 		})
 		c.Cookie(&fiber.Cookie{
 			Name:   "user",
@@ -99,28 +99,53 @@ func NewUser(c *fiber.Ctx) error {
 	user := new(User)
 	err := fiberutils.ParseBody(c, user)
 	user.Password, err = passwordhashing.HashPassword(user.Password)
-	db.Create(&user)
+	db.Create(&user) //db. is from gorm package which automatically create a query statement either create, update, delete or view //&-get reference from user
 	c.JSON(user)
 	return err
 }
 
 func UpdateUser(c *fiber.Ctx) error {
+	userToken := c.Locals("user").(*jwt.Token)
+	claims := userToken.Claims.(jwt.MapClaims)
+	userO := (claims["user"].(map[string]interface{}))
+
 	db := database.DBConn
 	user := new(User)
 	err := fiberutils.ParseBody(c, user)
-	user.Password, err = passwordhashing.HashPassword(user.Password)
-	db.Model(&user).Where("id = ?", user.Id).Updates(user)
-	c.JSON(user)
+
+	if userO["role"] == "admin" || fmt.Sprintf("%v", userO["id"]) == fmt.Sprintf("%d", user.Id) {
+		user.Password, err = passwordhashing.HashPassword(user.Password)
+		db.Model(&user).Where("id = ?", user.Id).Updates(user)
+		c.JSON(user)
+	} else {
+		c.JSON(fiberutils.Message{
+			Message: "No permission to update",
+			Status:  "failed",
+		})
+	}
+
 	return err
 }
 
 func DeleteUser(c *fiber.Ctx) error {
+	userToken := c.Locals("user").(*jwt.Token)
+	claims := userToken.Claims.(jwt.MapClaims)
+	userO := (claims["user"].(map[string]interface{}))
+
 	db := database.DBConn
 	id := c.Params("id")
 	user := new(User)
 	db.First(&user, id)
-	db.Delete(&user)
-	c.JSON(user)
+
+	if userO["role"] == "admin" || fmt.Sprintf("%v", userO["id"]) == fmt.Sprintf("%d", user.Id) {
+		db.Delete(&user)
+		c.JSON(user)
+	} else {
+		c.JSON(fiberutils.Message{
+			Message: "No permission to delete",
+			Status:  "failed",
+		})
+	}
 	return nil
 }
